requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
httpx==0.24.1
python-dotenv==1.0.0
gradio==3.40.1
Pillow==10.0.0
requests==2.31.0
google-cloud-firestore==2.11.0
google-cloud-storage